updated gitignore and updatedSquare2.c

+++ bookExercises/chapter7/updatedSquare2.c
@@ -3,19 +3,19 @@
 int main(void)
 {
 
-    short int i, n;
+    long int i, n;
 
     printf("This program prints a table of squares.\n");
 
     printf("Enter number of entries in table: ");
-    scanf("%hd", &n);
+    scanf("%ld", &n);
 
     for (i = 1; i <= n; i++){
-        if ((short)(i*i) < 0){
-            printf("the lowest number who is also the same type when squared is %hd\n", --i);
+        if ((long)(i*i) < 0){
+            printf("the lowest number who is also the same type when squared is %ld\n", --i);
             break;
         }
-        printf("%10d%10d\n", i, i * i);
+        printf("%10ld%10ld\n", i, i * i);
 
     }
